Fixed PR comments
