+++ packages/experimental/Avatar/src/useAvatar.ts
@@ -2,6 +2,7 @@ import { JSAvatarProps, AvatarInfo, JSAvatarState, AvatarColors } from './JSAvat
 import { ImageProps, ImageSourcePropType } from 'react-native';
 import { PresenceBadgeProps } from '@fluentui-react-native/badge';
 import { titles } from './titles';
+import { getHashCodeWeb } from './getHashCode';
 /**
  * Re-usable hook for FURN Avatar.
  * This hook configures Avatar props and state for FURN Avatar.
@@ -33,7 +34,7 @@ export const useAvatar = (props: JSAvatarProps): AvatarInfo => {
   };
 
   const _initials = initials || getInitials(name);
-  const avatarColorsIdx = getHashCode(idForColor ?? name ?? '') % AvatarColors.length;
+  const avatarColorsIdx = getHashCodeWeb(idForColor ?? name ?? '') % AvatarColors.length;
   const _avatarColor = avatarColor === 'colorful' ? AvatarColors[avatarColorsIdx] : avatarColor;
 
   return {
@@ -96,21 +97,4 @@ export const validateAlphabeticalCharacters = (name: string): boolean => {
  */
 export const removeTitlesFromName = (words: string[]): string[] => {
   return words.filter((word) => !titles.has(word));
-};
-
-/**
- * A function which returns hash code for the avatar color
- *
- * @param str is idForColor or name
- * @returns hash code of the color
- */
-const getHashCode = (str: string): number => {
-  let hashCode = 0;
-  for (let len: number = str.length - 1; len >= 0; len--) {
-    const ch = str.charCodeAt(len);
-    const shift = len % 8;
-    hashCode ^= (ch << shift) + (ch >> (8 - shift)); // eslint-disable-line no-bitwise
-  }
-
-  return hashCode;
 };
